requirements.txt
streamlit
pymupdf
openai
langchain
faiss-cpu
python-dotenv
